+++ src/pages/conference/_lazy-load.ts
@@ -2,4 +2,4 @@ import Loadable from '@loadable/component'
 
 export const Preview = Loadable(() => import('./_preview'))
 export const Venue = Loadable(() => import('./_venue'))
-export const NoticableLogin = Loadable(() => import('./_noticable-login'))
+export const Subscribe = Loadable(() => import('./_subscribe'))
