--- src/pages/conference/_map.tsx
@@ -1,14 +1,0 @@
-import React from 'react'
-import styled from 'styled-components'
-import { localize } from 'components/localization'
-import { Flex } from 'components/containers'
-import { Header } from 'components/elements'
-
-const Wrapper = styled.div``
-const Card = styled(Flex)``
-
-const Component = () => {
-    return <></>
-}
-
-export default Component
