+++ src/pages/conference/index.tsx
@@ -3,7 +3,7 @@ import Hero from './_hero'
 import Login from './_login'
 import Events from './_events'
 import Speakers from './_speakers'
-import { NoticableLogin, Preview, Venue } from './_lazy-load'
+import { Subscribe, Preview, Venue } from './_lazy-load'
 import { SEO } from 'components/containers'
 import Layout from 'components/layout/layout'
 import { WithIntl } from 'components/localization'
@@ -21,7 +21,7 @@ const Conferencies = () => {
             <Preview />
             <Speakers />
             <Venue />
-            <NoticableLogin />
+            <Subscribe />
         </Layout>
     )
 }
